Handle Juniper authgroup correctly.
In AnyConnect, the server explicitly tells us which group is selected and
we set form->authgroup_selection accordingly. But we're abusing this for
Juniper too, to allow users to use the --authgroup option for setting the
realm. Fix it to automatically set form->authgroup_selection based on the
existing setting of vpninfo->authgroup (which is automatically updated
in when the user sets it, too).

This should fix endless loops in the KDE auth dialog, and the "snapback"
problem when changing it results in a new form being presented, with
the option suddenly changed back to the default value.

Signed-off-by: David Woodhouse <dwmw2@infradead.org>