Support for multiple objects per PKCS#11 url
It's possible for multiple object to have the same PKCS#11 url,
which causes problems when selecting certificates, since it's not
possible to say which of the certificates to choose.

This patch checks if multiple certificates are available for the
same URL and allows the user to select one of them. If only one
certificate is available, it will be used automatically, thus preserving
the previous functionality.

Signed-off-by: Elias Norberg <elias@aisle.se>