Do not split fortinet cookies
Fortinet uses a single SVPNCOOKIE for authentication (which is base64
encoded), so sometimes it can end in a = or == for padding.
internal_split_cookies assumes that the cookie string is of the format
NAME=VALUE and tries to split it on the equal signs, which corrupts the
Fortinet cookie encoding. Fortinet only uses a single cookie anyway, so
skip splitting and just add the cookie directly. This allows connecting
in my local testing.

Signed-off-by: Jacob Garber <jacob.garber@calian.com>