GlobalProtect: Add external browser with CAS support
In order to handle 2FA login pages an external browser is necessary
until webkitgtk supports those methods out out the box.
Therefore add an option to make use of external browser depending
on default-browser setting. In order to return the callback data
from browser to openconnect a helper is necessary which registers
a MimeType in the system.

Signed-off-by: Jan-Michael Brummer <jan-michael.brummer1@volkswagen.de>

Fixes: https://gitlab.com/openconnect/openconnect/-/issues/651